Defines new package to arangodb
