+++ arangodb-extension/src/main/java/org/jnosql/artemis/arangodb/document/CouchbaseTemplate.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.jnosql.artemis.couchbase.document;
+package org.jnosql.artemis.arangodb.document;
 
 
 import com.couchbase.client.java.document.json.JsonObject;
