+++ arangodb-extension/src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -12,4 +12,4 @@
 #
 #   Otavio Santana
 #
-org.jnosql.artemis.couchbase.document.CouchbaseExtension
+CouchbaseExtension
