+++ arangodb-extension/src/test/java/org/jnosql/artemis/arangodb/document/CouchbaseExtensionTest.java
@@ -12,8 +12,26 @@
  *
  *   Otavio Santana
  */
-package org.jnosql.artemis.couchbase.document;
+package org.jnosql.artemis.arangodb.document;
 
+import org.junit.Test;
+import org.junit.runner.RunWith;
 
-public interface PersonRepositoryAsync extends CouchbaseRepositoryAsync<Person, String> {
+import javax.inject.Inject;
+
+@RunWith(WeldJUnit4Runner.class)
+public class CouchbaseExtensionTest {
+
+    @Inject
+    private PersonRepositoryAsync personRepositoryAsync;
+
+    @Inject
+    private PersonRepository personRepository;
+
+    @Test
+    public void shouldSaveOrientDB() {
+        Person person = new Person("Ada", 10);
+        personRepository.deleteById(person.getName());
+        personRepositoryAsync.deleteById(person.getName());
+    }
 }
