+++ arangodb-extension/src/test/java/org/jnosql/artemis/arangodb/document/DefaultCouchbaseTemplateAsyncTest.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.jnosql.artemis.couchbase.document;
+package org.jnosql.artemis.arangodb.document;
 
 import com.couchbase.client.java.document.json.JsonObject;
 import com.couchbase.client.java.query.Statement;
