+++ arangodb-extension/src/test/java/org/jnosql/artemis/arangodb/document/MockProducer.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.jnosql.artemis.couchbase.document;
+package org.jnosql.artemis.arangodb.document;
 
 
 import org.jnosql.diana.api.document.Document;
