+++ arangodb-extension/src/test/java/org/jnosql/artemis/arangodb/document/PersonRepositoryAsync.java
@@ -12,20 +12,8 @@
  *
  *   Otavio Santana
  */
-package org.jnosql.artemis.couchbase.document;
+package org.jnosql.artemis.arangodb.document;
 
-import org.junit.runners.BlockJUnit4ClassRunner;
-import org.junit.runners.model.InitializationError;
 
-public class WeldJUnit4Runner extends BlockJUnit4ClassRunner {
-
-    public WeldJUnit4Runner(Class<Object> clazz) throws InitializationError {
-        super(clazz);
-    }
-
-    @Override
-    protected Object createTest() {
-        final Class<?> test = getTestClass().getJavaClass();
-        return WeldContext.INSTANCE.getBean(test);
-    }
+public interface PersonRepositoryAsync extends CouchbaseRepositoryAsync<Person, String> {
 }
